+++ actions/crash-course.ts
@@ -3,10 +3,9 @@ import { Resend } from 'resend';
 import { PrismaClient } from '@prisma/client';
 import { CourseRegisterSchema } from '@/schemas';
 import * as z from 'zod';
-
+import RegCrashCourse from '@/lib/reg-crashcourse';
 
 const resend = new Resend(process.env.RESEND_API_KEY)
-const chatLink = `https://wa.me/+2348163453826`;
 
 export const CrashCourse = async (values: z.infer<typeof CourseRegisterSchema>) => {
     const validatedFields = CourseRegisterSchema.safeParse(values);
@@ -23,6 +22,7 @@ export const CrashCourse = async (values: z.infer<typeof CourseRegisterSchema>) 
     if (!validatedFields.success) {
         return { error: 'Invalid fields!'};
     }
+    const {course,email,expectations,firstname,lastname,occupation} = validatedFields.data
 
     /**This constant gets the validated fields in registration form and sends the submitted information to 
      * the provided TO_EMAIL_ADDRESS.
@@ -30,12 +30,12 @@ export const CrashCourse = async (values: z.infer<typeof CourseRegisterSchema>) 
      */
     try {
         const emailContent = `
-        Firstname: ${validatedFields.data.firstname}
-        Lastname: ${validatedFields.data.lastname}
-        Email: ${validatedFields.data.email}
-        Course: ${validatedFields.data.course}
-        Occupation: ${validatedFields.data.occupation}
-        Expectations: ${validatedFields.data.expectations}
+        Firstname: ${firstname}
+        Lastname: ${lastname}
+        Email: ${email}
+        Course: ${course}
+        Occupation: ${occupation}
+        Expectations: ${expectations}
         `;
 
         await resend.emails.send({
@@ -50,23 +50,22 @@ export const CrashCourse = async (values: z.infer<typeof CourseRegisterSchema>) 
          * It uses the users email address and takes the first name of the user and 
          * sends the reply message
          */
+
         await resend.emails.send({
             from: process.env.FROM_EMAIL_ADDRESS!,
             to: validatedFields.data.email,
             subject: 'Crash Course Registration',
-            text: `Dear ${validatedFields.data.firstname}
-            Thank you for registering for the Crash Course! We've received your registration.
-            `,
+            react: RegCrashCourse({courseName: course,email, firstname})
 
         });
 
             const registration = await prisma.courseRegistration.create({ 
                 data: {
-                  firstname: validatedFields.data.firstname,
-                  lastname: validatedFields.data.lastname,
-                  email: validatedFields.data.email,
-                  occupation: validatedFields.data.occupation,
-                  course: validatedFields.data.course
+                  firstname,
+                  lastname,
+                  email,
+                  occupation,
+                  course
                 }
             })
         
@@ -75,7 +74,10 @@ export const CrashCourse = async (values: z.infer<typeof CourseRegisterSchema>) 
          * form must have checked for errors and found non or found some
          * the CATCH catches the error and displays it to the user else it displays the SUCCESS message.
          */
-        return { success: 'Registration successful! An acknowledgment email has been sent to you!' };
+        return { 
+            success: 'Registration successful! An acknowledgment email has been sent to you!',
+            data: registration
+        };
     } catch (error) {
         console.error('Error sending email:', error);
         return {error: 'An error occurred while submitting. Please try again. Or check your internet connection.'}