+++ app/sitemap.ts
@@ -13,8 +13,6 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
         }
   })
 
-
-
   return [
       {
         url: `${process.env.NEXT_PUBLIC_URL}/about`,