+++ components/pages/about/our-values.tsx
@@ -12,29 +12,15 @@ export default function OurValues() {
 			className='min-h-fit pt-16 border-b'
 		>
 			<div className='grid md:grid-cols-2 gap-16 items-center mx-auto w-full justify-center'>
-				<div className='grid grid-cols-2 gap-5 md:text-2xl text-md text-wrap'>
+				<div className='grid grid-cols-2 gap-5 md:text-xl text-sm text-wrap'>
 					<div className='text-2xl font-bold pt-12 col-span-2'>{`Our Values`}</div>
 					<div className='col-span-2'>
-						<span className=''>{`We plan to lead the charge of technological innovation and solutions in africa while improving consumers lives.`}</span>
+						<span className=''>{`We strive to be the foundation for our students' success and empowerment. By providing a supportive and engaging learning environment, we help our students build the skills and confidence they need to thrive. Our commitment is to empower individuals to fulfill their goals and make positive contributions to their communities.`}</span>
 					</div>
-					<div className=' '>
-						<span className='font-normal text-black flex flex-col'>{`Transparency`}</span>
-					</div>
-					<div className=''>
-						<span className=''>{`We are the solution `}</span>
-					</div>
-					<div className=''>
-						<span className=''>{`Be relevant `}</span>
-					</div>
-					<div className=''>
-						<span className=''>{`Connectivity is our reality `}</span>
-					</div>
-					<div className=''>
-						<span className=''>{`Focus `}</span>
-					</div>
+					
 				</div>
 
-				<div className='md:text-2xl text-md text-wrap  '>
+				<div className='md:text-xl text-sm text-wrap  '>
 					<div className='w-full h-full rounded-xl overflow-clip my-3'>
 						<Image
 							src='/innovation.jpg'